Implemented screen resizing and scrolling ground tiles.
Details:
All operations that deal with the screen, such as rendering to the screen, or storing coordinates, are done in a fixed size surface. This surface is simply scaled up to fit the window once the surface is ready to be shown. This makes the code simpler as there's no need to actually change the coordinates of the objects on the screen. The user's mouse coordinates are scaled back down before being used for collision detection.

When the window is taller than the game screen, the bottom of the window is filled with the same color as the ground, which makes the extra space join seamlessly with the bottom of the game screen. The top of the window is then filled with the game's sky color. All extra space is filled with blue when the window is wider than the game screen.

The bottom of the screen is filled with ground tiles moving leftwards, which gives the illusion of leftward movement.