Several changes and new features.
Changes:
The movement of the ground tiles now compensate for framerate changes, so the speed will remain the same regardless of framerate.

New Features:
Added the bird.
The bird wiggles using a sin wave while waiting for the player's first tap.
Implemented movement physics for the bird, including falling from gravity and jumping on a tap from the player.
The bird changes the angle it's facing in proportion to its movement speed.

Note that collision detection is not implemented yet, so the bird will fall/jump out of bounds if tapped too frequently or not fast enough.