Partially implemented collision detection.
The bird now has a hit-box that resizes every frame to fit the size of the bird. Ceiling and ground detection has also been added. When the bottom side of the bird's hit-box is colliding with the top of the screen, the game will prevent the bird from moving any higher. And when the bird collides with the ground, the death animation plays. The death animation causes the bird to jump once, then fall through the ground and out of the screen. Note that the lose screen is not implemented yet, so the game will do nothing after the bird falls out of the screen. Pipes are not added yet, so currently hitting the ground is the only way to die.