Added pipes.
Details:
Pipes have been added, and collision detection with them has also been implemented. Now there are two ways to die, hitting a pipe and hitting the ground. The pipe which the bird hit flashes to indicate the collision. The scene also stops scrolling once the bird dies. The results screen is still not added yet.