Improved collision detection, bug fixes, and code clean-up.
Changes:
Improved the accuracy of collision detection with the pipes and ground.
The centered rotation system was rewritten to use a much better method that does not cause inaccuracy in mask collision detection, and also makes code less confusing.
The padding color of the window has been changed from cyan to black.
Rewrote the huge block of spaghetti code in the collision detection section of "main.py".
Wrapped up several blocks of code in functions to reduce the code clutter in the game loop.
Removed the yellow rectangle that is filled into the bottom padding of the window.
Debug mode is toggled when the U key is pressed. Debug mode displays hit-boxes to the objects on the screen, and also shows the collide mask of the bird.
Rainbow mode is toggled when the J key is pressed. This makes the padding of the window cycle through randomly generated colors. Note that this effect is not noticeable when the window is adjusted to a ratio that fits the content perfectly, or when the window is at its default size.
Full-screen mode is toggled when the F11 function key is pressed.

Notes:
The method of toggling debug mode and rainbow mode will be changed in the future to something more hidden and complicated than a single keyboard button press.

An FPS counter will be added to debug mode in the future.

The game-over screen is *still* not implemented yet, oops.