Added FPS counter and key sequence detecter.
Changes:
An FPS counter has been added to debug mode. The FPS is calculated by finding the average of the delay between the last 10 frames. Due to this, the counter will remain blank for 10 frames after activating debug mode.

A class for detecting key sequences has been written to simplify adding easter-eggs :)

The activation method of rainbow mode has been changed from pressing the J key to entering the Konami code (↑,↑,↓,↓,←,→,←,→,b,a). The activation will only succeed if the keys are pressed in the correct order with no more than 0.5 seconds of delay in-between.

The activation method of debug mode has been changed from pressing the U key to entering the magic word (xyzzy).