Partially implemented toast notifications.
Details:
A toast notifier class has been added. When called with a string passed as the message content, it soft-wraps the text until it is small enough to fit in the hard-coded frame width (which is currently set at 320px). Then it slides the notification into the screen with a smooth Euler deceleration. Notification icons have not been drawn yet, so currently the place where the icon goes in the notification is just a black box. The close button is also not implemented yet, so currently there's no way to dismiss a notification once it is shown.

For testing and demonstration purposes, pressing the T key in-game brings up a mock toast notification filled with dummy text. This key can be pressed multiple times to show multiple notifications on the screen. This function will probably be removed in the next commit.