Finished implementing toast notifications.
Details:
Toast notifications now have a close button that can be clicked to dismiss them. The close button is also not drawn yet, so currently it uses a black box as a placeholder for the close button. When a toast notification is dismissed, all notifications (if any) below the one dismissed will snap up to fill the space. The demo toast notification summoned by the 'T' key has not been removed yet. Toast notifications with real messages will be added in the next commit, and the demo one will be removed.