Bug fix: Fixed a bug in 'Global.py' which causes an unhandled exception when run on non-win32 OSes.
Details:
'Global.py' only successfully imports 'windll' if it is run on win32 (obviously), but later code in the file attempts to reference the variable 'windll' without taking into consideration that it might not exist. Therefore, the game will raise an exception and crash if run on a non-win32 OS. I have fixed this by changing the code which references 'windll' to handle 'NameError' if it is raised. Note that the Windows binary release and MSI installer have not been re-compiled with this fix, as they can only run on Windows anyway ¯\_(ツ)_/¯.